backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.1.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
django-extensions>=3.2.0
django-ninja>=0.22.2
anthropic>=0.8.1 